update vendordeps, reorganization and cleanup
