+++ src/main/java/frc/robot/Configs.java
@@ -3,8 +3,6 @@ package frc.robot;
 import com.revrobotics.spark.config.SparkMaxConfig;
 import com.revrobotics.spark.config.ClosedLoopConfig.FeedbackSensor;
 import com.revrobotics.spark.config.SparkBaseConfig.IdleMode;
-
-import edu.wpi.first.wpilibj.motorcontrol.Spark;
 import frc.robot.Constants.ModuleConstants;
 
 public final class Configs {
