--- src/main/java/frc/robot/MAXSwerve-Java-2025-main-Imported.code-workspace
@@ -1,7 +1,0 @@
-{
-    "folders": [
-        {
-            "path": "../../../../.."
-        }
-    ]
-}
